Create class SymbolArray.
SymbolArray logically represents a sequence of symbols. The array is
split up into chunks and each chunk can have a bunch of unused bytes at
the end. The unused bytes are skipped over in O(1) time when moving
back/forward in the sequence of symbols.

This representation will make it possible to implement an efficient
parallel version of the re-pair algorithm.