More parallelism in replacePairs.
Allow two chunks to be processed in parallel even if X==Y and the first
two symbols in the second chunk is equal to X and the last symbol in the
first chunk is also equal to X.

This improves compression speed especially if a file contains mostly
zeros. It makes the code quite a bit more complicated though.