Faster integer division in index2Pos().
Speed up the integer divisions in PosIndex::index2Pos() but using the
fact that the same denominator is used many times. It is therefore
possible to replace division by r with multiplication by
2^(64+log2(r))/r and shifting the result. This requires a 128 bit
integer data type though, (__int128 in gcc), so there is also a standard
C++ implementation which can be enabled at compile time.