Improve tree pruning threshold.
Always use entropy when deciding what tree nodes to prune, since entropy
can be meaningfully compared to the size required to encode the decision
tree. Also change the default pruning threshold to 4.1, which is roughly
the expected size increase in bytes when adding one more node to the
decision tree.