Better iterative algorithm for building decision tree.
Only apply a subset of the positions to the tree StatsCollectorNodes. If
a node has a small enough estimated error, replace it with its
corresponding predicate node and create new child nodes.

This strategy means that the tree does not grow one level at a time.

Currently this method is slower than the old method because the "small
enough estimated error" strategy is too conservative.