Use permutation generators when building the decision tree.
Use a permutation generator to directly compute the indices to sample,
instead of testing all indices to determine whether they should be
sampled.

This is faster for sparse enough sampling, but has a larger constant
overhead because of less regular memory access pattern. Therefore it is
slower for dense sampling.